refactor: refine user data handling and API responses
Updates the API to return user data wrapped in a `UserSimpleResource` for consistency.

Improves data retrieval by including the user's role and additional fields in the `UserRepository`.

+++ app/Services/UserService.php
@@ -67,7 +67,7 @@ class UserService
     public function deleteUser(string $id) {
         $user = $this->getUserById($id);
 
-        if(Storage::disk("public")->exists($user->photo)) {
+        if(!empty($user->photo) && Storage::disk("public")->exists($user->photo)) {
             Storage::disk("public")->delete($user->photo);
         }
